Rename diag-option && move parsing to DiagnosticsEngine
